requirements.txt
requests
cryptography
python-dotenv
pytest
pytest-html
